+++ packages/gamut-styles/CHANGELOG.md
@@ -1,3 +1,25 @@
+## [7.0.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@6.0.0...@codecademy/gamut-styles@7.0.0) (2020-09-01)
+
+
+### ⚠ BREAKING CHANGES
+
+* **gamut-styles:** deprecate legacy JS color variables  (#940)
+
+### Other Changes
+
+* **gamut-styles:** deprecate legacy JS color variables  ([#940](https://github.com/Codecademy/client-modules/issues/940)) ([903de62](https://github.com/Codecademy/client-modules/commit/903de62bf3086e3d51ba6dec36cc77ca482ba52f))
+
+## [6.0.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@5.1.2...@codecademy/gamut-styles@6.0.0) (2020-08-28)
+
+
+### ⚠ BREAKING CHANGES
+
+* Swapped default font family to Apercu (#935)
+
+### ✨ Features
+
+* Swapped default font family to Apercu ([#935](https://github.com/Codecademy/client-modules/issues/935)) ([ce20167](https://github.com/Codecademy/client-modules/commit/ce20167524c2302cff218261d10a81eb3deb41c0)), closes [#943](https://github.com/Codecademy/client-modules/issues/943)
+
 ### [5.1.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@5.1.1...@codecademy/gamut-styles@5.1.2) (2020-08-24)